POC using Go + xsltproc to pursue XLST 3.0 approach
xlstproc is not easily available within Docker. Overall the JSON translation approach for the XML data is not a good approach. Need to focus on handling the XML as-is with xpath access. I don't think transforming the data set is going fit into the scope here. If it does - Go is not the language.