Successful POC of calculating word & section count using eCFR API endpoints
Utilize the summary API to get section details for a given agency. Use the returned details to query for the full XML content. Do a simple word count on the content returned. This is an imperfect count as there is some markup in the content - but relatively it is accurate.

Rate limiting will prevent this approach from scaling, but it atleast narrows down how word and section count can be calculated. Need to capture the entire XML dataset so that calculations can be done without eCFR APIs.

Examples:
- http://localhost:3000/word-count/northeast-interstate-low-level-radioactive-waste-commission
- http://localhost:3000/word-count/appalachian-regional-commission