Delete XSLT import programs
This motivation behind this approach was to avoid pulling majority of data from the API. In theory, creating a robust XSLT translation would enable utilizing the bulk export files to populate the database. Seems like this still would be a good approach - but it would take me a long time to get right.

Instead I am going to move forward by slowly pulling the data I need from the API. Rate limiting will be the main obstacle with this approach.