[PINE-316] Utilize simple Node scripts to crawl CFR bulk data API and collect archive links
The intent was to collect archive links and either download and persist them inside of this Node script - but overall this will be better handled server side.

I would rather have a single DB connection point (i.e. do not have connections in these scripts + in the server), and I don't see the benefit in collecting the URLs in a script to just pass them to the server for download.

Unlike the ECFR files, the CFR archives have several volumes within each title, so there needs to be some additional iterating once the archive is downloaded.